# Conflicts:
#	baby_p2p-framework-code/baby_p2p-framework-code.iml
#	baby_p2p-framework-common/baby_p2p-framework-common.iml
#	baby_p2p-framework-oss/baby_p2p-framework-oss.iml
#	baby_p2p-framework-safe/baby_p2p-framework-safe.iml

+++ baby_p2p-framework-code/src/main/java/com/xkkj/code/pojo/Bid.java
@@ -38,10 +38,10 @@ public class Bid implements Serializable {
     private String borrowTitle;
 
     @ApiModelProperty(value = "投标金额")
-    private Long bidAmount;
+    private Integer bidAmount;
 
     @ApiModelProperty(value = "投标利息")
-    private Long bidInterest;
+    private Integer bidInterest;
 
     @ApiModelProperty(value = "年化率")
     private Integer yearRate;
